Move enhance, toArray out of lib, remove babel from devDeps
