+++ src/createExposedMachine.js
@@ -1,4 +1,4 @@
-import enhance from "./lib/enhance";
+import enhance from "./enhance";
 import expose from "./plugins/expose";
 
 export default enhance(expose);
