+++ src/index.js
@@ -1,5 +1,5 @@
 import createMachine from "./createMachine";
 import createExposedMachine from "./createExposedMachine";
-import enhance from "./lib/enhance";
+import enhance from "./enhance";
 
 export { enhance, createMachine, createExposedMachine };
